correstion de tous ce qui concerne produits

+++ database/migrations/2025_07_10_160133_create_mouvement_produits_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('mouvement_produits', function (Blueprint $table) {
             $table->id('mouvementProd_id');
-            $table->integer('produit_id');
+            $table->unsignedBigInteger('produit_id');
             $table->date('date');
             $table->string('origine')->nullable(); 
             $table->integer('quantite_commandee');
